+++ packages/tailwindcss/src/compat/config/create-compat-config.ts
@@ -14,6 +14,10 @@ export function createCompatConfig(cssTheme: Theme): UserConfig {
       colors: ({ theme }) => theme('color', {}),
 
       extend: {
+        fontSize: ({ theme }) => ({
+          ...theme('text', {}),
+        }),
+
         boxShadow: ({ theme }) => ({
           ...theme('shadow', {}),
         }),